product_service/requirements.txt
fastapi
psycopg2-binary
sqlalchemy
uvicorn
python-jose[cryptography]
httpx
aio_pika
redis
asyncpg